requirements.txt
streamlit
SpeechRecognition
pyaudio
python-dotenv
requests
together
Pillow
arabic_reshaper
python-bidi
googletrans
